// types/Arqueo.d.ts
interface Arqueo {
  id: stringi;
  nombres: string;
  documento: string;
  sucursal: string;
  supervisor: string;
  puntodeventa: string;
  latitud: string;
  longitud: string;

  ventabruta: string;
  baseefectivo: string;
  cartera: string;
  totalingreso: string;

  chancesabonado: string;
  chancespreimpresos: string;
  premiospagados: string;
  efectivocajafuerte: string;
  tirillarecaudo: string;
  totalegresos: string;

  totalbilletes: string;
  totalmonedas: string;
  totalarqueo: string;
  sobrantefaltante: string;

  totalbilletescaja: string;
  totalmonedascaja: string;
  totalpremioscaja: string;
  total: string;

  rollos_bnet: string;
  rollos_fisicos: string;
  diferencia: string;

  nombre_juego: string;
  cantidad_bnet: string;
  cantidad_fisicos: string;
  cantidad_faltante: string;
  cantidad_tiquete: string;
  descargado: string;

  nombre_juego2: string;
  cantidad_bnet2: string;
  cantidad_fisicos2: string;
  cantidad_faltante2: string;
  cantidad_tiquete2: string;
  descargado2: string;

  nombre_juego3: string;
  cantidad_bnet3: string;
  cantidad_fisicos3: string;
  cantidad_faltante3: string;
  cantidad_tiquete3: string;
  descargado3: string;

  nombre_juego4: string;
  cantidad_bnet4: string;
  cantidad_fisicos4: string;
  cantidad_faltante4: string;
  cantidad_tiquete4: string;
  descargado4: string;

  nombre_juego5: string;
  cantidad_bnet5: string;
  cantidad_fisicos5: string;
  cantidad_faltante5: string;
  cantidad_tiquete5: string;
  descargado5: string;

  nombre_juego6: string;
  cantidad_bnet6: string;
  cantidad_fisicos6: string;
  cantidad_faltante6: string;
  cantidad_tiquete6: string;
  descargado6: string;

  nombre_juego7: string;
  cantidad_bnet7: string;
  cantidad_fisicos7: string;
  cantidad_faltante7: string;
  cantidad_tiquete7: string;
  descargado7: string;

  totaldescargados: string;
  totalvalor: string;

  requisito1: string;
  requisito2: string;
  requisito3: string;
  requisito4: string;
  requisito5: string;
  requisito6: string;
  requisito7: string;
  requisito8: string;
  requisito9: string;
  requisito10: string;
  requisito11: string;
  requisito12: string;
  requisito13: string;
  requisito14: string;
  requisito15: string;
  requisito16: string;
  requisito17: string;
  requisito18: string;
  requisito19: string;
  requisito20: string;
  requisito21: string;
  requisito22: string;
  requisito23: string;
  requisito24: string;
  requisito25: string;
  requisito26: string;
  requisito27: string;
  requisito28: string;
  requisito29: string;
  requisito30: string;
  requisito31: string;
  requisito33: string;
  requisito32: string;
  requisito34: string;
  requisito35: string;

  nombre_observacion: string;
  firma_colocadora?: string;
  firma_auditoria?: string;
  imagen_observacion?: string;
}

export type Arqueos = Arqueo[];